Implement model resources
