+++ modules/v1/setup/resources/UserProfileResource.php
@@ -1,0 +1,17 @@
+<?php
+
+
+namespace app\modules\v1\setup\resources;
+
+
+use app\models\UserProfile;
+
+/**
+ * Class UserProfileResource
+ *
+ * @package app\modules\v1\setup\resources
+ */
+class UserProfileResource extends UserProfile
+{
+
+}
